+++ app/backend/src/tests/test_app.py
@@ -1,7 +1,6 @@
 import pytest
 
 from couchers.server import create_main_server, create_media_server
-from tests.test_fixtures import testconfig  # noqa
 
 
 @pytest.fixture(autouse=True)
