+++ app/backend/src/tests/test_auth.py
@@ -21,18 +21,13 @@ from couchers.models import (
     UserSession,
 )
 from couchers.proto import api_pb2, auth_pb2
-from tests.test_fixtures import (  # noqa
-    PushCollector,
+from tests.test_fixtures import (
     api_session,
     auth_api_session,
-    db,
     email_fields,
-    fast_passwords,
     generate_user,
     mock_notification_email,
-    push_collector,
     real_api_session,
-    testconfig,
 )
