+++ app/backend/src/tests/test_bugs.py
@@ -7,7 +7,10 @@ from google.protobuf import empty_pb2
 from couchers.config import config
 from couchers.crypto import random_hex
 from couchers.proto import bugs_pb2
-from tests.test_fixtures import bugs_session, db, generate_user, testconfig  # noqa
+from tests.test_fixtures import (
+    bugs_session,
+    generate_user,
+)
 
 
 @pytest.fixture(autouse=True)
