+++ app/backend/src/tests/test_discussions.py
@@ -5,15 +5,11 @@ from couchers.db import session_scope
 from couchers.proto import discussions_pb2, notifications_pb2, threads_pb2
 from couchers.utils import now, to_aware_datetime
 from tests.test_communities import create_community, create_group
-from tests.test_fixtures import (  # noqa
-    PushCollector,
-    db,
+from tests.test_fixtures import (
     discussions_session,
     generate_user,
     notifications_session,
     process_jobs,
-    push_collector,
-    testconfig,
     threads_session,
 )
