+++ app/backend/src/tests/test_donations.py
@@ -12,7 +12,11 @@ from couchers.jobs.handlers import update_badges
 from couchers.models import DonationInitiation, DonationType, Invoice, InvoiceType, User, UserBadge
 from couchers.proto import donations_pb2
 from couchers.proto.google.api import httpbody_pb2
-from tests.test_fixtures import db, donations_session, generate_user, real_stripe_session, testconfig  # noqa
+from tests.test_fixtures import (
+    donations_session,
+    generate_user,
+    real_stripe_session,
+)
 
 
 @pytest.fixture(autouse=True)
