+++ app/backend/src/tests/test_editor.py
@@ -11,7 +11,10 @@ from couchers.models import (
     Volunteer,
 )
 from couchers.proto import editor_pb2
-from tests.test_fixtures import db, generate_user, real_editor_session, testconfig  # noqa
+from tests.test_fixtures import (
+    generate_user,
+    real_editor_session,
+)
 
 
 @pytest.fixture(autouse=True)
