+++ app/backend/src/tests/test_jail.py
@@ -7,19 +7,14 @@ from couchers.models import users
 from couchers.proto import admin_pb2, api_pb2, jail_pb2
 from couchers.servicers import jail as servicers_jail
 from couchers.utils import create_coordinate, to_aware_datetime
-from tests.test_fixtures import (  # noqa
-    PushCollector,
-    db,
+from tests.test_fixtures import (
     email_fields,
-    fast_passwords,
     generate_user,
     mock_notification_email,
-    push_collector,
     real_account_session,
     real_admin_session,
     real_api_session,
     real_jail_session,
-    testconfig,
 )
