+++ app/backend/src/tests/test_media.py
@@ -8,7 +8,11 @@ from couchers.crypto import b64decode, random_hex
 from couchers.db import session_scope
 from couchers.models import InitiatedUpload, Upload
 from couchers.proto import media_pb2
-from tests.test_fixtures import api_session, db, generate_user, media_session, testconfig  # noqa
+from tests.test_fixtures import (
+    api_session,
+    generate_user,
+    media_session,
+)
 
 
 @pytest.fixture(autouse=True)
