+++ app/backend/src/tests/test_references.py
@@ -25,20 +25,15 @@ from couchers.models import (
 from couchers.moderation.utils import create_moderation
 from couchers.proto import conversations_pb2, references_pb2, requests_pb2
 from couchers.utils import create_coordinate, now, to_aware_datetime, today
-from tests.test_fixtures import (  # noqa
-    PushCollector,
+from tests.test_fixtures import (
     account_session,
-    db,
     email_fields,
     generate_user,
     make_friends,
     make_user_block,
     mock_notification_email,
-    moderator,
-    push_collector,
     references_session,
     requests_session,
-    testconfig,
 )
 from tests.test_requests import valid_request_text
