+++ app/backend/src/tests/test_reporting.py
@@ -5,7 +5,10 @@ from sqlalchemy import select
 from couchers.db import session_scope
 from couchers.models import ContentReport
 from couchers.proto import reporting_pb2
-from tests.test_fixtures import db, generate_user, reporting_session, testconfig  # noqa
+from tests.test_fixtures import (
+    generate_user,
+    reporting_session,
+)
 
 
 @pytest.fixture(autouse=True)
