+++ app/backend/src/tests/test_requests.py
@@ -25,18 +25,13 @@ from couchers.proto.internal import unsubscribe_pb2
 from couchers.rate_limits.definitions import RATE_LIMIT_DEFINITIONS, RATE_LIMIT_HOURS
 from couchers.templates.v2 import v2date
 from couchers.utils import create_coordinate, now, today
-from tests.test_fixtures import (  # noqa
-    PushCollector,
+from tests.test_fixtures import (
     api_session,
     auth_api_session,
-    db,
     email_fields,
     generate_user,
     mock_notification_email,
-    moderator,
-    push_collector,
     requests_session,
-    testconfig,
 )
