+++ app/backend/src/tests/test_threads.py
@@ -8,7 +8,10 @@ from couchers.db import session_scope
 from couchers.models import Thread
 from couchers.proto import threads_pb2
 from couchers.servicers.threads import pack_thread_id
-from tests.test_fixtures import db, generate_user, testconfig, threads_session  # noqa
+from tests.test_fixtures import (
+    generate_user,
+    threads_session,
+)
 
 
 @pytest.fixture(autouse=True)
