+++ app/backend/src/tests/test_verification.py
@@ -12,16 +12,11 @@ from couchers.db import session_scope
 from couchers.models import SMS, User
 from couchers.proto import account_pb2, api_pb2
 from couchers.utils import now
-from tests.test_fixtures import (  # noqa
-    PushCollector,
+from tests.test_fixtures import (
     account_session,
     api_session,
-    db,
     generate_user,
-    notifications_session,
     process_jobs,
-    push_collector,
-    testconfig,
 )
